requirements.txt
gspread
google-api-python-client
google-auth-oauthlib
selenium
webdriver-manager
python-dotenv
google-generativeai
schedule
